Add support for blending operations to CGContext (#2222)
This pull request adds support for blending to CGContext. There are a number of
caveats.

* `kCGBlendModeClear` is not supported.
* The *Source In*, *Destination In*, *Source Out*, *Destination Atop*, and
  *Source Copy* blend modes cannot be used properly without transparency layers.
  Their use per-primitive will result in unusual artifacts.
* `kCGBlendModePlusDarker` is not supported (and was not supported in the
original Cairo implementation).

Some blend modes (the ones without Source/Destination) in the name require a
full buffer read-back and compose operation. This is unavoidable.

This change unfortunately _does_ include a rewrite of the CGD2D rendering
stack. Instead of adding yet another nesting level of command lists and layers,
I switched it to a scaffold/commit stack. Depending on various conditions,
operations will be queued up, started in order, and committed in reverse order.
A great many of those operations involve creating command lists and layers and
blending them or otherwise composing them.

There is a not-insignificant performance penalty in certain cases:

|                                    before             after   delta
| CoreText__CTLineDrawComplete       742.51            716.11   -3.5546% faster |
| CoreText__CTLineDrawGroup         292,909           307,828    5.0934% slower |
| CoreText__CTLineDrawSingle         233.43            280.05   19.9746% slower |

times in μs.

While 19% seems like a lot, it's only 47μs/operation; an application running at
60fps requires its rendering operations complete in ~16000μs. It's not
_significant_, but it is a cost taken per operation.

I believe there is opportunity for improvement; specifically, I had to
compromise on allocations to achieve the desired effects of polymorphism in the
new render operation stack.

Fixes #1389.