Support custom fonts in Xaml through _compatibleFamilyName (#2297)
Our implementation of supporting user-supplied fonts could not propagate to XAML because we cannot modify the system's font collection and instead have a separate font collection where we store user-supplied fonts. XAML can't access this separate font collection, and instead has to pull the font from the file directly. We already had a private method _compatibleFamilyName which returned the family name required for XAML as CTFontGetFamilyName would return different strings for some fonts, so this has been modified to return the string which XAML will use to get the specified font from its file.

This will be in format @"ms-appx///[Absolute Path To File].ttf#FamilyName". This however requires that the family name requested from the font be used for that specific font, as families can have multiple files. For example, WinObjC-Bold.ttf has family name WinObjC as does WinObjC.ttf, but getting _compatibleFamilyName from a regular WinObjC font will not return the correct path to construct a Bold font, so one should be careful to create the font UIFont with required attributes before trying to convert to XAML.

We will get the path to the file by saving the CGDataConsumer which contains the CFURL it is registered with in the IDWriteFontFile we create. Then, when we go to get the _compatibleFamilyName we retrieve the CGDataConsumer from which we retrieve the CFURL, pull out the path, and return the correct value.

This can be manually confirmed in XamlCatalog through the Custom Text Editor, which now uses our WinObjC font (the character 'A' is represented by a solid black circle).

Fixes #2273