Unify all WinObjC Versioning
This changes simplifies and unifies all versioning to be based on Git Tags.
This removes the need for the version file and no longer requires master packages
to be built with a version override (though they still can).

vcxproj, csproj, and VSIX projects all get automatically versioned in a concsistent
way and have more metadata like the Copyright and description correctly filled out.