Redo Any CPU build multiplexing (#2422)
* Redo Any CPU build multiplexing

The problem with the existing implementation is two fold:

1. MSBuild isn't as smart as I thought.
2. I'm not as smart as I thought.

more seriously the two issues were:

1. GetPackageContents was not correctly invoking ResolveReferences
2. Global property differences between .nuproj build multiplexing, test multiplexing, and .nuproj -> .nuproj project references was triggering incorrect rebuilds.

In the new mechanism, the build of all frameworks will be complete before GetPackageContents is invoked on that framework. This is because GetPackageContents for .nuproj projects now depends on ResolveReferences. This means that inside of GetPackageContents (which invokes a new instance of MSBuild due to an extra global property), no build resolution is needed so for .vcxproj projects, it does NOT depend on ResolveReference. This makes GetPackageContents a little more lightweight but requires all package file inference to not depend on building. Additionally, when multiplexing any "multiplex only" global properties are removed before invoking project reference builds to uniformly invoke them.