Add support for NSUncaughtExceptionHandler (#2432)
This commit introduces support for Foundation's last line of defense
against uncaught exceptions. The application is being torn down, and
there's nothing a consumer can do to stop it.

They can now at least log something.

This change cannot be automatically tested, and it cannot be tested with
an attached debugger.  This was tested manually (via MessageBox) on x86.

There's no user32 or MessageBox on ARM, and since we're in an exception
handler it would be dangerous to try to create any UI or interact too
much with COM and WinRT.

Since named exception catch is already broken on ARM (#352), this likely
does not work. No tests (including file writing tests) succeeded.

Fixes #2373.