Rudimentary share dialog from Social kit (#2314)
Add rudimentary Social kit share view controller

This change introduces a rudimentary sharing dialog for Facebook, provided by the Social kit. Callers can create a SLComposeViewController and add a link or image to it. When presented, the view controller will pop up a dialog box allowing users to add a message and then post the content to Facebook.

The user must already be logged in to Facebook for this to work. The sharing dialog does not do login on its own. It retrieves an existing Facebook access token from the Windows password vault.

This implementation is unfinished, so don't be surprised to see parts of the code that aren't complete. There are a lot of features of the sharing dialog that have yet to be added.