Switch from projections to C++/WinRT (#2382)
Switch from projections to C++/WinRT

This change switches some Framework code from using projections to using C++/WinRT when interacting with WinRT objects. This is beneficial because C++/WinRT is lighter-weight than the projections. They're header-only, so we don't have to link with the ObjCUWP_* binaries or load zillions of Objective-C classes.

In this change, UIKit has been switched away from using the XAML projections, along with a couple of other smaller frameworks. Eventually none of the frameworks should be using projections, and the projections will exist solely for use by application code.