Restore dependency walking to get minset of binaries
This works by removing extra items from ReferenceCopyLocalPaths by walking the target
binary's dependency (and any implicit dependencies from References) and keying off of
a special piece of metadata added only by WinObjC binaries. NOTE: this requires that
dependencies and references all be transitive. In other words by depending on a package,
and producing a package, a consumer of the second package also depends on the first. This
creates an end result where all binaries are trying to be copied into the final output and
the final output has extra references added to it since the package can't distinguish between
a "true consumer" and a "transitive consumer".

Additionally clean up sbclang.targets and other packaging items to do with linking.