Check validity of m_phase before use
Projects are allowed to have no build phase to associate to m_phase.
For example, if there are no resources to build then there would be no
resource build phase enabling m_phase to remain NULL.

Some portions of vsimporter rely on m_phase to be NULL, so we must do a
check when we actually want to access it.