ClangCompile: stop freshening object files all the time
The file freshness detector in ClangCompile had the following logic:

If the project file is older than the object file (almost always true),
- OR -
If the object's build command line was the same as the last build,

we would update the modification time of the object. In the first case,
we would probably update the modtime of every object produced by the
project.

We were successfully skipping compilation, but the modtime freshening
was triggering a re-link of all downstream consumers. This would happen
up to three times during a parallel build.

It looks like there's no value in updating the modification times any
more, so we can get rid of entire swaths of code that do that.