tests: consolidate all properties into ut-build, make it like sdk-build (#2913)
The critical difference between `sdk-build` and `ut-build` is that `sdk-build` was intended to be included _instead of_ `Microsoft.Cpp.Common`, which allowed it to change a whole host of settings before the C++ infrastructure was initialized. This commit switches `ut-build` to the same pattern. It also removes an obscene number of identical properties and simplifies the test projects quite a bit.