Fixed generics declarations for collection subclasses.
Generics declarations of NSMutableArray, NSMutableDictionary, NSMutableOrderedSet, and NSMutableSet were not being passed up to superclass.

This meant that calling generic methods declared in the superclass on a subclass object were not recognized by the compiler to return a declared generic object type.

All protocol declarations that are being replaced with generics declarations are already declared by the respective superclass, and thus the list of implemented protocols does not change. Also removed unecessary protocol declarations on NSMutableIndexSet to match the other collection subclasses.

Also added missing generics declaration to NSCountedSet (subclass of NSMutableSet).